Please fix ComponentNext -> ComponentNest
This example seems to have been misnamed. It should have been ComponentNest. I can't rename the folder to fix this for you from the github ui: Example 14 on http://angular2-first-look.azurewebsites.net/ and obviously there are more files to change. It was an issue raised by one of your users on StackBlitz's issues. Sorry I've no time to raise the full PR for you.